requirements.txt
--extra-index-url https://download.pytorch.org/whl/nightly/cpu
--pre
pandas
llama-index
python-dotenv
psycopg2-binary
llama-index-vector-stores-postgres
torch
torchvision
torchaudio
--index-url https://pypi.org/simple
sentence-transformers
llama-index-embeddings-huggingface
grpcio
grpcio-tools
protobuf